feat: split kindasafe into no_std read crate and kindasafe_init
Separate kindasafe into two crates so that signal-handler-path code
can depend on the lightweight read primitives without pulling in libc:

- kindasafe: #![no_std] crate with zero production dependencies,
  containing only the naked-asm memory read primitives (u64, slice,
  str, fs_0x10) and CrashPoint types.

- kindasafe_init: new crate depending on kindasafe + libc + spin,
  containing all initialization and signal handler orchestration
  (init, crash_handler, fallback chaining, sanity_check).
