+++ MANIFEST.in
@@ -2,8 +2,8 @@ include Cargo.toml
 include Cargo.lock
 include pyroscope_ffi/python/rust/Cargo.toml
 include pyroscope_ffi/ruby/ext/rbspy/Cargo.toml
-include kit/kindasafe/Cargo.toml
 recursive-include src *.rs
 recursive-include pyroscope_ffi/python/rust/src/ *.rs
 recursive-include pyroscope_ffi/ruby/ext/rbspy/src/ *.rs
-recursive-include kit/kindasafe/src/ *.rs
+recursive-include kit/ *.rs
+recursive-include kit/ Cargo.toml
