+++ kit/kindasafe/Cargo.toml
@@ -4,8 +4,8 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-libc =  {version = "0.2.182"}
-spin =  {version="0.10.0"}
+
 [dev-dependencies]
 anyhow = "=1.0.102"
 libc = "=0.2.182"
+kindasafe_init = { path = "../kindasafe_init" }
