+++ kit/kindasafe_init/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "kindasafe_init"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+kindasafe = { path = "../kindasafe" }
+libc = { version = "0.2.182" }
+spin = { version = "0.10.0" }
+
+[dev-dependencies]
+anyhow = "=1.0.102"
